feat: creating MarkdownText and UrlLauncher
