+++ catalyst_voices/packages/internal/catalyst_voices_shared/lib/src/catalyst_voices_shared.dart
@@ -34,5 +34,6 @@ export 'utils/active_aware.dart';
 export 'utils/date_time_ext.dart';
 export 'utils/future_ext.dart';
 export 'utils/iterable_ext.dart';
+export 'utils/launch_url_mixin.dart';
 export 'utils/lockable.dart';
 export 'utils/typedefs.dart';
