+++ catalyst_voices/packages/internal/catalyst_voices_shared/lib/src/utils/launch_url_mixin.dart
@@ -1,0 +1,9 @@
+import 'package:url_launcher/url_launcher.dart';
+
+mixin LaunchUrlMixin {
+  Future<void> launchHrefUrl(Uri url) async {
+    if (!await launchUrl(url)) {
+      throw Exception('Could not launch $url');
+    }
+  }
+}
