+++ catalyst_voices/packages/internal/catalyst_voices_shared/pubspec.yaml
@@ -28,6 +28,7 @@ dependencies:
   logging: ^1.2.0
   shared_preferences: ^2.3.3
   uuid: ^4.5.1
+  url_launcher: ^6.2.2
   web: ^1.1.0
 
 dev_dependencies:
